Relocate Splash page to its own package.
