+++ ui/lib/routing/widget/harmony_router_delegate.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:harmony/app_shell/app_shell.dart';
-import 'package:harmony/spec/spec.dart';
+import 'package:harmony/splash/splash.dart';
 
 import '../config/app_route_state.dart';
 import '../config/harmony_page.dart';
@@ -13,8 +12,6 @@ final pageToPathMap = <HarmonyPage, HarmonyRoutePath>{
   HarmonyPage.welcome: WelcomeRotuePath(),
   HarmonyPage.splash: SplashRoutePath(),
 };
-
-final _assetName = 'assets/Harmony.svg';
 
 class HarmonyRouterDelegate extends RouterDelegate<HarmonyRoutePath>
     with PopNavigatorRouterDelegateMixin<HarmonyRoutePath>, ChangeNotifier {
@@ -60,7 +57,7 @@ class HarmonyRouterDelegate extends RouterDelegate<HarmonyRoutePath>
           ),
         if (appState.selectedPage == HarmonyPage.splash)
           MaterialPage(
-            child: _TempSplashPage(
+            child: SplashPage(
               onComplete: () {
                 appState.selectedPage = HarmonyPage.welcome;
               },
@@ -112,58 +109,5 @@ class HarmonyRouterDelegate extends RouterDelegate<HarmonyRoutePath>
 
   void _setPage(HarmonyPage page) {
     appState.selectedPage = page;
-  }
-}
-
-class _TempSplashPage extends StatefulWidget {
-  _TempSplashPage({required this.onComplete});
-
-  final VoidCallback onComplete;
-  @override
-  __TempSplashPageState createState() => __TempSplashPageState();
-}
-
-// TODO(#32): Create a Splash page for users's loading the application
-class __TempSplashPageState extends State<_TempSplashPage>
-    with TickerProviderStateMixin {
-  late final AnimationController _controller = AnimationController(
-    duration: const Duration(seconds: 3),
-    vsync: this,
-  )
-    ..addStatusListener((status) {
-      if (status == AnimationStatus.completed) widget.onComplete();
-    })
-    ..forward();
-
-  late final Animation<double> _animation = CurvedAnimation(
-    parent: _controller,
-    curve: Curves.easeOutCubic,
-  ).drive(Tween(begin: 0, end: 4.25));
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final logoSvg = SvgPicture.asset(
-      _assetName,
-      semanticsLabel: 'Harmony Logo',
-      fit: BoxFit.scaleDown,
-      height: Sizing.serverSideBarWidth,
-    );
-    return Scaffold(
-      body: Center(
-        child: RotationTransition(
-          turns: _animation,
-          child: TextButton(
-            onPressed: widget.onComplete,
-            child: logoSvg,
-          ),
-        ),
-      ),
-    );
   }
 }
