+++ ui/lib/splash/splash.dart
@@ -1,0 +1,1 @@
+export 'page/splash_page.dart';
